requirements.txt
mccabe (0.4.0)
numpy (1.11.0)
Pillow (3.2.0)
pip (8.1.1)
pyflakes (1.0.0)
pygame (1.9.2a0)
python-gnupg (0.3.8)
PyYAML (3.11)
setuptools (20.6.7)
wheel (0.29.0)
